% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_solve.R
\name{solution_status}
\alias{solution_status}
\title{Return the solution status.}
\usage{
solution_status(result)
}
\arguments{
\item{result}{a cbc solution object}
}
\value{
a string being either "optimal", "infeasible" or "unbounded"
}
\description{
Return the solution status.
}
