% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_solve.R
\name{column_solution}
\alias{column_solution}
\title{Return the column solution}
\usage{
column_solution(result)
}
\arguments{
\item{result}{a cbc solution object}
}
\value{
numeric vector. One component per column in the
         order of the column vector.
}
\description{
Return the column solution
}
