% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_solve.R
\name{objective_value}
\alias{objective_value}
\title{Return the objective value of a solution}
\usage{
objective_value(result)
}
\arguments{
\item{result}{a cbc solution object}
}
\value{
a single number
}
\description{
Return the objective value of a solution
}

